[Silver III] Title: 2×n 타일링, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ C++17/백준/Silver/11726. 2×n 타일링/2×n 타일링.cc
@@ -4,10 +4,10 @@
 using namespace std;
 
 int main() {
-	long long n;
+	int n;
 	cin >> n;
 
-	vector<long long> cases(n + 1);
+	vector<int> cases(n + 1);
 	cases[0] = 1;
 	cases[1] = 1;
